requirements.txt
# Editable install with no version control (agent==0.0.1)
-e /Users/javcab/myralph/raphael
annotated-types==0.7.0
anyio==4.12.1
backoff==2.2.1
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.1
coloredlogs==15.0.1
distro==1.9.0
exceptiongroup==1.3.1
flatbuffers==25.2.10
gnupg==2.3.1
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
kittens==0.1.2
kitty==1.2.0
langchain-core==1.2.7
langgraph==1.0.6
langgraph-checkpoint==4.0.0
langgraph-cli==0.4.11
langgraph-prebuilt==1.0.6
langgraph-sdk==0.3.3
langsmith==0.6.4
mem0ai==1.0.3
mpmath==1.3.0
numpy==2.2.6
onnx==1.18.0
onnxruntime==1.22.1
onnxruntime-tools==1.7.0
openai==2.21.0
orjson==3.11.5
ormsgpack==1.12.2
packaging==25.0
portalocker==3.2.0
posthog==7.9.0
protobuf==5.29.6
psutil==7.0.0
py-cpuinfo==9.0.0
py3nvml==0.2.7
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
qdrant-client==1.16.2
requests==2.32.4
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.46
sympy==1.14.0
tenacity==9.1.2
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.14.1
urllib3==2.5.0
uuid_utils==0.14.0
xmltodict==0.14.2
xxhash==3.6.0
zstandard==0.25.0
